Change type of finish reason from enum to enum | string to allow for new enums by providers and to not throw a runtime exception
