+++ src/providers/ai21/types.ts
@@ -1,0 +1,5 @@
+
+export enum AI21_FINISH_REASON {
+  stop = 'stop',
+  length = 'length'
+}
