+++ src/providers/anthropic/types.ts
@@ -1,0 +1,7 @@
+
+export enum ANTHROPIC_STOP_REASON {
+  max_tokens = 'max_tokens',
+  stop_sequence = 'stop_sequence',
+  tool_use = 'tool_use',
+  end_turn = 'end_turn'
+}
